2009-02-12	Joel Sherrill <joel.sherrill@oarcorp.com>
	* sapi/include/rtems/config.h, score/include/rtems/score/thread.h,
	score/src/threadidlebody.c: Change prototype of IDLE thread to
	consistently return void * and take a uintptr_t argument.

+++ cpukit/sapi/include/rtems/config.h
@@ -128,7 +128,7 @@ typedef struct {
   /** This element points to the BSP's optional idle task which may override
    *  the default one provided with RTEMS.
    */
-  Thread                       (*idle_task)( uint32_t );
+  Thread                       (*idle_task)( uintptr_t );
 
   /** This field specifies the size of the IDLE task's stack.  If less than or
    *  equal to the minimum stack size, then the IDLE task will have the minimum
